ngo frontend/requirements.txt
Flask>=2.0
flask-cors
reportlab
python-dotenv
# razorpay is optional — only needed if USE_RAZORPAY=1
razorpay
